Upload/View de documentos dependentes - Alterações de responsividade

+++ application/controllers/Dependentes.php
@@ -3,7 +3,23 @@
 class Dependentes extends MY_Controller {
 
     public function novo() {
-        $this->load->view('novo_dependente');
+        try {
+            $id = $this->input->get('id');
+            if (empty($id)) {
+                throw new Exception('Não é possível acessar esta página.');
+            }
+
+            $this->load->model('associado');
+            $this->load->model('dao/associadoDAO', 'ad');
+            $this->associado->setId($id);
+
+            $this->load->template('novo_dependente', [
+                'active' => 'associados',
+                'associado' => $this->ad->listarMenos($this->associado)
+            ]);
+        } catch (Exception $e) {
+            $this->error($e);
+        }
     }
 
     public function editar($id) {
@@ -12,9 +28,10 @@ class Dependentes extends MY_Controller {
             $this->load->model('dao/dependenteDAO', 'dd');
             $this->dependente->setId($id);
             $data = [
+                'active' => 'associados',
                 'dependente' => $this->dd->listarPorId($this->dependente),
             ];
-            $this->load->view('editar_dependente', $data);
+            $this->load->template('editar_dependente', $data);
         } catch (Exception $e) {
             $this->error($e);
         }
@@ -34,46 +51,45 @@ class Dependentes extends MY_Controller {
         }
     }
 
-    public function salvar($id, $idDependente = null) {
+    public function salvar($idDependente = null) {
         try {
             $this->load->model('associado');
             $this->load->model('dependente');
             $this->load->model('dao/dependenteDAO', 'dd');
-            $this->associado->setId($id);
-            $this->dependente->setAssociado($this->associado);
             $this->dependente->setNome($this->input->post('nome'));
             $this->dependente->setNascimento($this->input->post('nascimento'), true);
             $this->dependente->setCpf($this->input->post('cpf'));
             $this->dependente->setParentesco($this->input->post('parentesco'));
+            $associadoID = null;
             if (empty($idDependente)) {
+                $this->associado->setId($this->input->post('associado-id'));
+                $associadoID = $this->associado->getId();
+                $this->dependente->setAssociado($this->associado);
                 $this->dd->inserir($this->dependente);
             } else {
                 $this->dependente->setId($idDependente);
                 $this->dd->alterar($this->dependente);
+                $associadoID = $this->dd->listarPorId($this->dependente)->getAssociado()->getId();
             }
-            print json_encode(['status' => true]);
+            redirect('associados/editar/' . $associadoID . '/#dependentes');
         } catch (Exception $e) {
-            print json_encode([
-                'status' => false,
-                'error' => $e->getMessage(),
-            ]);
+            $this->error($e);
         }
     }
 
-    public function excluir($id) {
+    public function excluir($id = null) {
         try {
+            $associadoID = $this->input->get('id');
+            if (empty($id) || empty($associadoID)) {
+                throw new Exception('Não foi possível acessar esta página.');
+            }
             $this->load->model('dependente');
             $this->load->model('dao/dependenteDAO', 'dd');
             $this->dependente->setId($id);
             $this->dd->delete($this->dependente);
-            print json_encode([
-                'status' => true,
-            ]);
+            redirect('associados/editar/' . $associadoID . '/#dependentes');
         } catch (Exception $e) {
-            print json_encode([
-                'status' => false,
-                'error' => $e->getMessage(),
-            ]);
+            $this->error($e);
         }
     }
